requirements.txt
Flask
discord.py
requests